refactor: make tool names easier to interpret
